requirements.txt
azure-devops~=7.1.0
typer~=0.13.0
textual~=0.85.2